Convert XC models to interface
