+++ subprojects/build-adapter-xcode/src/main/java/dev/nokee/xcode/XCProjectLoader.java
@@ -58,6 +58,6 @@ public final class XCProjectLoader implements XCLoader<XCProject, XCProjectRefer
 		val schemeNames = builder.build();
 
 		// TODO: Add support for implicit scheme: xcodebuild -list -project `getLocation()` -json
-		return new XCProject(reference.getName(), reference.getLocation(), targets, schemeNames, pbxproj, fileReferencesLoader);
+		return new DefaultXCProject(reference.getName(), reference.getLocation(), targets, schemeNames, pbxproj, fileReferencesLoader);
 	}
 }
